requirements.txt
asgiref==3.10.0
Django==5.2.7
django-admin-sortable2==2.2.8
django-crispy-forms==2.4
pillow==12.0.0
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.11.0
gunicorn==20.1.0